# services/inventory/models/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package models

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Inventory struct {
	ID        pgtype.UUID
	OwnerID   pgtype.UUID
	Name      string
	Category  string
	Color     string
	Size      string
	Brand     string
	Location  string
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}

type User struct {
	ID           pgtype.UUID
	Email        string
	FirstName    string
	LastName     string
	PasswordHash string
	CreatedAt    pgtype.Timestamp
	UpdatedAt    pgtype.Timestamp
}